carselling/requirements.txt
Django==4.0.3
django-filter==21.1
django-widget-tweaks==1.4.12
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1